% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot_Functions.r
\name{calc_KMcurve_pvalues}
\alias{calc_KMcurve_pvalues}
\title{Tests if there is a difference between two or more survival curves using the G-rho family of tests, or for a single curve against a known alternative.}
\usage{
calc_KMcurve_pvalues(survdiff, digits = 3)
}
\arguments{
\item{survdiff}{object generated by the survdiff() function from survival package}

\item{digits}{number of digits in the p-value to report}
}
\value{
a p-value
}
\description{
Tests if there is a difference between two or more survival curves using the G-rho family of tests, or for a single curve against a known alternative.
}
\examples{
TBD <- c("")
}
