% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE_Cleaning_Utility_Functions.r
\name{collapseDuplicates}
\alias{collapseDuplicates}
\title{For each duplicate USI, collapse the columns' information so that it will be seperated by a semi-colon.}
\usage{
collapseDuplicates(df, ID.column, duplicate)
}
\arguments{
\item{df}{the data.frame}

\item{ID.column}{the name of the column with sample IDs}

\item{duplicate}{boolean if sample ID is duplicated or not}
}
\value{
data.frame
}
\description{
For each duplicate USI, collapse the columns' information so that it will be seperated by a semi-colon.
}
\examples{
my_data <- data.frame(Sample=paste0("s",1:20), mutation=rep(c("Yes","No"), length.out=20))
collapseDuplicates(my_data, ID.column = "Sample", duplicate=c("s2","s5"))

}
