% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{collapseRows}
\alias{collapseRows}
\title{Collapse duplicate rows so that all information is retain,optionally unique, and separated by a semi-colon.}
\usage{
collapseRows(col, uniq = FALSE, split = FALSE, sep = "")
}
\arguments{
\item{col}{colname}

\item{uniq}{boolean}

\item{split}{boolean}

\item{sep}{delimiter}
}
\value{
string vector
}
\description{
Collapse duplicate rows so that all information is retain,optionally unique, and separated by a semi-colon.
}
\examples{
metadata <- data.frame(mutation = rep("yes","no", length.out = 10))
collapseRows(col = "mutation")
}
