% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{colorCodes_aheatmap}
\alias{colorCodes_aheatmap}
\title{Create a color codes}
\usage{
colorCodes_aheatmap(df, random = FALSE)
}
\arguments{
\item{df}{cleaned clinical characteristics, with patient IDs as rownames}

\item{random}{boolean. should colors be selected by random sampling.}
}
\value{
list
}
\description{
Create a color codes
}
\examples{
ex <- c('TBD')
}
