% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{colorDends}
\alias{colorDends}
\title{Generate a dendrogram with branches colored by groups}
\usage{
colorDends(hclustObject, colorCodes, group, textsize)
}
\arguments{
\item{hclustObject}{hclust object is the ward or other clustering matrix from hclust()}

\item{colorCodes}{colorcodes is a vector with group status="color" format}

\item{group}{group is the character vector with groups status, one for each patient (sample_id as names)}

\item{textsize}{textsize is a numeric vector of length 2 with size for 1) the labels of axes, and 2) sample_id on leaves}
}
\value{
plot
}
\description{
Generate a dendrogram with branches colored by groups
}
\examples{
ex <- c("TBD")
}
