% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{colorDends_Groups}
\alias{colorDends_Groups}
\title{Function for Grouping and Splitting Trees}
\usage{
colorDends_Groups(
  dendrogram,
  phenovector,
  k = NULL,
  h = NULL,
  branchCol = NULL,
  colorcodes
)
}
\arguments{
\item{dendrogram}{a dendrogram object}

\item{phenovector}{character vector with groups status, one for each patient (sample_id as names)}

\item{k}{number of clusters}

\item{h}{height to cut the dendrogram}

\item{branchCol}{branchCol is a vector of length K, with the colors for the branches.}

\item{colorcodes}{color codes is a named character vector for the known groups (eg c(pos="red, neg="blue)) for the leaves.}
}
\value{
list
}
\description{
Function for Grouping and Splitting Trees
}
\examples{
ex <- c("TBD")
}
