% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{create_HA_Labs_Hmap}
\alias{create_HA_Labs_Hmap}
\title{Create ComplexHeatmap annotation object}
\usage{
create_HA_Labs_Hmap(
  expn,
  geneList,
  CDE,
  cols,
  goi = NULL,
  cc = NULL,
  colorbar.height = 5
)
}
\arguments{
\item{expn}{the normalized expression values with genes as rownames}

\item{geneList}{a character vector}

\item{CDE}{clinical data. has column called sample_id}

\item{cols}{character vector of column names in CDE}

\item{goi}{genes of interest to label on the Heatmap. Character vector of gene symbols}

\item{cc}{color codes as a named character vector.}

\item{colorbar.height}{numeric heigh in cm}
}
\value{
list
}
\description{
Create ComplexHeatmap annotation object
}
\examples{
ex <- c('TBD')
}
