% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeGSEA-package.R
\docType{package}
\name{DeGSEA-package}
\alias{DeGSEA}
\alias{DeGSEA-package}
\title{DeGSEA: Differential Expression and GSEA workflows along with Vizualization Functions}
\description{
A package of utility scripts for correlating RNA-seq transcript expression data to clinical covariates in the TARGET project.
}
\author{
\strong{Maintainer}: Jenny Smith \email{jennyl.smith12@gmail.com} (\href{https://orcid.org/0000-0003-0402-2779}{ORCID})

}
\keyword{internal}
