% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{dge_dendrograms}
\alias{dge_dendrograms}
\title{Create dendrogram from DGE object}
\usage{
dge_dendrograms(
  expnData,
  pheno,
  method,
  genelist = NULL,
  add.count = 0.01,
  percent = 0.01,
  filterTopGenes = FALSE,
  createDGE = TRUE,
  log = FALSE
)
}
\arguments{
\item{expnData}{normalized expression data or count data, patient IDs are column names and rows are genes.}

\item{pheno}{a character vector with patient IDs as names, and the status for each in each group (eg pos,neg)}

\item{method}{cluster method}

\item{genelist}{a character vector with the genes of interest}

\item{add.count}{psuedocount value for log2 transformation}

\item{percent}{is the \% of samples in the input expn matrix that must express a gene at 1 CPM. Filter to remove low count genes.}

\item{filterTopGenes}{boolean. If TRUE filter top 1000 most varied genes.}

\item{createDGE}{boolean. TRUE if raw counts are input.}

\item{log}{boolean. Yes/No for log2 transformation.}
}
\value{
list
}
\description{
Create dendrogram from DGE object
}
\examples{

ex <- c('TBD')
}
