% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{extract_DEGs}
\alias{extract_DEGs}
\title{Title}
\usage{
extract_DEGs(
  twoGroups_DEGs.res,
  filter = FALSE,
  goi = NULL,
  anno = FALSE,
  geneLevel = FALSE
)
}
\arguments{
\item{twoGroups_DEGs.res}{output of twoGroups_DEGs()}

\item{filter}{boolean remove duplicate genes}

\item{goi}{genes of interest}

\item{anno}{boolean for gene annots}

\item{geneLevel}{boolean for gene annots or transcript annots}
}
\value{
dataframe
}
\description{
Title
}
\examples{
ex <- c("TBD")

}
