% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGE_GSEA_GSVA_Function.r
\name{gage_from_pipeline}
\alias{gage_from_pipeline}
\title{Wrapper Function to run GAGE GSEA on the output for the DEGs pipeline}
\usage{
gage_from_pipeline(
  twoGroups_DEGs.res,
  type,
  geneset = NULL,
  method = "voom",
  min_size = 15,
  ...
)
}
\arguments{
\item{twoGroups_DEGs.res}{output from twoGroups_DEGs()}

\item{type}{type is either limma differential expression results (FC) or expression matrix of normalized counts (expn)}

\item{geneset}{optional list gene-sets}

\item{method}{voom or trend}

\item{min_size}{min size of gene-sets}

\item{...}{others}
}
\value{
list
}
\description{
Wrapper Function to run GAGE GSEA on the output for the DEGs pipeline
}
\examples{
ex <- c("TBD")
}
