% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGE_GSEA_GSVA_Function.r
\name{GAGE_GSA}
\alias{GAGE_GSA}
\title{GAGE gene-set enrichment analysis}
\usage{
GAGE_GSA(df, type, geneset = NULL, ref = NULL, samp = NULL, min_size = 15)
}
\arguments{
\item{df}{df has genes symbols as rownames and patients as columns with normalized expression data}

\item{type}{type is either limma differential expression results (FC) or expression matrix of normalized counts (expn)}

\item{geneset}{optional list gene-sets}

\item{ref}{column numbers for reference samples}

\item{samp}{column numbers for samples in condition of interest}

\item{min_size}{min size of gene-sets}
}
\value{
list
}
\description{
GAGE gene-set enrichment analysis
}
\examples{
ex <- c("TBD")
}
