% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{gene_protein_anno}
\alias{gene_protein_anno}
\title{Annote transmembrane, subcellular location, and signal peptide genes, and ADC drug information.}
\usage{
gene_protein_anno(
  df,
  gene.name.col = "gene",
  ids2symbols = NULL,
  mart.37 = NULL,
  mart.38 = NULL,
  makeQuery = TRUE,
  attempts = 5
)
}
\arguments{
\item{df}{dataframe with genes as rows. May be expression dataset or DE genes list.}

\item{gene.name.col}{character vector of column name if any dataframe has an alternative column name for gene symbols (eg BCL2,TP53)}

\item{ids2symbols}{an be NULL, NA, or a file path. NULL is for BCCA id mapping + backwards compatibility.}

\item{mart.37}{NULL or provide biomaRt object for GRCh37}

\item{mart.38}{NULL or provide biomaRt object for GRCh38}

\item{makeQuery}{boolean}

\item{attempts}{numeric}
}
\value{
data.frame
}
\description{
Annote transmembrane, subcellular location, and signal peptide genes, and ADC drug information.
}
\examples{
ex <- c('TBD')

}
