% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{getIDmap}
\alias{getIDmap}
\title{Create a gene ID map for ensembl gene_id, and transcript_ids}
\usage{
getIDmap(GTF)
}
\arguments{
\item{GTF}{a gtf file}
}
\value{
a dataframe
}
\description{
Create a gene ID map for ensembl gene_id, and transcript_ids
}
\examples{
gtf <- data.frame(V3 = "transcript", V9 = "attr1 red; attr2 blue; attr green")
getIDmap(gtf)
}
