% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGE_GSEA_GSVA_Function.r
\name{gsva_DE}
\alias{gsva_DE}
\title{DE analysis with the gsva matrix.}
\usage{
gsva_DE(gsva_matrix, clinData, col, p.value = 0.001)
}
\arguments{
\item{gsva_matrix}{output scores from gsva}

\item{clinData}{dataframe of clinical data with patients as rownames}

\item{col}{covariate column name}

\item{p.value}{p-value threshold}
}
\value{
list
}
\description{
DE analysis with the gsva matrix.
}
\examples{
ex <- c("TBD")
}
