% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot_Functions.r
\name{KM.plots}
\alias{KM.plots}
\title{Create Kaplan-Meier Plots}
\usage{
KM.plots(
  df,
  group_vars = NULL,
  column_names = list(OS = c("Overall.Survival.Time.in.Days", "OS.ID")),
  covariate,
  cc = NULL,
  riskTable = TRUE,
  f.levels = NULL,
  max.year = NULL,
  include.censored = TRUE
)
}
\arguments{
\item{df}{cleaned CDE with patient IDs as rownames.}

\item{group_vars}{character vector of a column name to dplyr::group_by() in dplyr.}

\item{column_names}{list with colnames OS for OS and EFS for "EFS" in the CDE}

\item{covariate}{character string for the column in CDE to be the explanatory variable}

\item{cc}{color codes with one color per group}

\item{riskTable}{boolean to generate a risk table under the KM plot}

\item{f.levels}{character vector of factor levels for the risk table}

\item{max.year}{year to calculate the statistics in the risk table (eg 5yrs after follow-up)}

\item{include.censored}{boolean to draw the tick marks on the curve to show censored points}
}
\value{
a grouped data.frame with the surv objects and the KM plots
}
\description{
Create Kaplan-Meier Plots
}
\examples{
my_cols <-  list("OS"=c("OS.time","OS.id"), "EFS"=c("EFS.time","EFS.id"))

}
