% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_MA_Plot_Function.r
\name{MA_plot}
\alias{MA_plot}
\title{MA plot}
\usage{
MA_plot(limma_eBayesFit, cut.off = 2.5, label.offset = 0.5)
}
\arguments{
\item{limma_eBayesFit}{a limma eBayesFit object}

\item{cut.off}{a numeric value of log2 fold-change for gene labels on points}

\item{label.offset}{a numeric value for nudge_y in ggrepel::geom_text_repel}
}
\value{
a ggplot object
}
\description{
MA plot
}
\examples{
output <- c('TBD')
}
