% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDE_Cleaning_Utility_Functions.r
\name{mergeMolCols}
\alias{mergeMolCols}
\title{merge molecular annotations side-by-side from the clinical data elements}
\usage{
mergeMolCols(col1, col2)
}
\arguments{
\item{col1}{the unquoted column name}

\item{col2}{the unquoted column name}
}
\value{
data.frame
}
\description{
merge molecular annotations side-by-side from the clinical data elements
}
\examples{
set.seed(1)
my_data <- data.frame(mutationA=sample(c("Yes","No"), size=20, replace=T), mutationB=sample(c("Yes","No"), size=20, replace=T))

}
