% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot_Functions.r
\name{outcome_table}
\alias{outcome_table}
\title{create a table with OS/EFS percent at specified time point}
\usage{
outcome_table(fit, time = 5, pvalues = NULL)
}
\arguments{
\item{fit}{the output of the survFit() function}

\item{time}{the time-point when to calculate the statistics (eg 5 yrs after follow-up)}

\item{pvalues}{optionally input p-values manually from survdiff() function}
}
\value{
a data.frame with the outcome statistics
}
\description{
create a table with OS/EFS percent at specified time point
}
\examples{
TBD <- c("")
}
