% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis_Function.r
\name{PCA}
\alias{PCA}
\title{Generate PCA plot}
\usage{
PCA(
  expnData,
  phenovector,
  title = "",
  round = TRUE,
  colorCodes = NULL,
  ntop = 500,
  PC3 = FALSE,
  GOI = NULL
)
}
\arguments{
\item{expnData}{raw counts (not normalized) has patient IDs as colnames and genes as rownames.}

\item{phenovector}{named character vector with patient ID = groupID}

\item{title}{character vector}

\item{round}{boolean}

\item{colorCodes}{named vector like c(groupID = 'red')}

\item{ntop}{number of most varied genes to use in PCA}

\item{PC3}{plot PC1 vs PC3 as well as PC1 vs PC2}

\item{GOI}{character vector of gene names/ gene id that match the expnData.}
}
\value{
list
}
\description{
Generate PCA plot
}
\examples{
ex <- c('TBD')
}
