% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis_Function.r
\name{pca_custom}
\alias{pca_custom}
\title{Custom PCA plot}
\usage{
pca_custom(
  expnData,
  CDE,
  fillCol,
  colorCol,
  colorCode = NULL,
  PC3 = FALSE,
  single.col.outline = FALSE,
  toHighlight = NULL,
  ellipse = FALSE
)
}
\arguments{
\item{expnData}{expn data is log2, normalized counts}

\item{CDE}{dataframe - has patients as rownames}

\item{fillCol}{character string of column name for fill colors}

\item{colorCol}{character string of column name for border colors}

\item{colorCode}{an option named vector of colors}

\item{PC3}{boolean to generate PC1 vs PC3 plot}

\item{single.col.outline}{boolean whether to susbet to dataframe for the border color}

\item{toHighlight}{if single.col.outline is true, Gives the factor to highlight with borders from the colorColumn}

\item{ellipse}{boolean to draw an ellipse based on fillCol column}
}
\value{
list of plots
}
\description{
Custom PCA plot
}
\examples{
mat <- matrix(runif(100,min = 0, max = 15), nrow = 10, ncol = 10)
metadata <- data.frame(mutation = rep("yes","no", length.out = 10))
pca_custom(mat, metadata, colorCol = "mutation")
}
