% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterAnalysis_Function.r
\name{plotPCoA}
\alias{plotPCoA}
\title{Multidimensional Scaling plot}
\usage{
plotPCoA(
  expnData,
  phenovector,
  title = "",
  colorCodes = NULL,
  geneList = NULL,
  size = 3
)
}
\arguments{
\item{expnData}{normalized counts, typically log2 scale.}

\item{phenovector}{named character vector of patient ID = groupID}

\item{title}{plot title}

\item{colorCodes}{named vector like groupID = 'red'}

\item{geneList}{character vector gene_names or gene_ids}

\item{size}{point size for the plot}
}
\value{
list
}
\description{
Multidimensional Scaling plot
}
\examples{
ex <- c('TBD')
}
