% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGE_GSEA_GSVA_Function.r
\name{read.gmt}
\alias{read.gmt}
\title{Read in a .gmt file and return a list of pathways}
\usage{
read.gmt(file)
}
\arguments{
\item{file}{file path to gmt}
}
\value{
list
}
\description{
Read in a .gmt file and return a list of pathways
}
\examples{
ex <- c("TBD")
}
