% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDupGenes_Function.r
\name{rmDups}
\alias{rmDups}
\title{Remove duplicate genes from an expression matrix}
\usage{
rmDups(count.matrix, ID.map, matrix.class = TRUE, rowname.GeneSym = TRUE)
}
\arguments{
\item{count.matrix}{matrix or dataframe with expression values. Genes as rownames, samples as columns}

\item{ID.map}{ID.map is a ensemble ID to gene symbol dataframe with colnames gene_id and gene_name.}

\item{matrix.class}{boolean to return a matrix or dataframe object}

\item{rowname.GeneSym}{return gene symbols as the rownames. If false, will have ensembl IDs as rownames}
}
\value{
an expression matrix or dataframe
}
\description{
Remove duplicate genes from an expression matrix
}
\examples{
ex <- c("TBD")
}
