% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmDupGenes_Function.r
\name{RPKM_to_TPM}
\alias{RPKM_to_TPM}
\title{Function for the TPM conversion.}
\usage{
RPKM_to_TPM(RPKM)
}
\arguments{
\item{RPKM}{expression matrix in RPKM}
}
\value{
matrix
}
\description{
Function for the TPM conversion.
}
\examples{
ex <- c("TBD")
}
