% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot_Functions.r
\name{SurvivalPlot}
\alias{SurvivalPlot}
\title{Survival plot without gridlines}
\usage{
SurvivalPlot(
  fit,
  LegendTitle,
  timeUnit,
  include.censored = TRUE,
  colors,
  pval = NULL,
  max.year = NULL
)
}
\arguments{
\item{fit}{the output of the survFit() function.}

\item{LegendTitle}{character vector which will be the plot legend title.}

\item{timeUnit}{character vector for the follow-up time in days, years, months etc.}

\item{include.censored}{boolean whether the add the ticks to indicate censored}

\item{colors}{character vector the same length as the number of groups.}

\item{pval}{string (charcter or numeric) of the p-value}

\item{max.year}{time (singe numeric) in years to end the plot at.}
}
\value{
a ggplot2::ggplot object
}
\description{
Survival plot without gridlines
}
\examples{
 TBD <- ("")
}
