% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survplot_Functions_2018.10.24.r
\name{SurvObjects}
\alias{SurvObjects}
\title{Generate a list of KM and Cox PH objects}
\usage{
SurvObjects(df, colNames, group, rho = 0, time = NULL, ref = NULL)
}
\arguments{
\item{df}{dataframe with clinical data.}

\item{colNames}{character vector (length 2) with the colnames of the time and event, in that order.}

\item{group}{column name as character vector of the explanatory variable to test, or numeric value of 1 for 1 line KM}

\item{rho}{1 or 0 or inbetween. 0=logrank test and 1=gehan-wilcoxon test}

\item{time}{optional if you need to convert days to years for example.}

\item{ref}{character indicating the group classification that is the reference}
}
\value{
a list with survfit objects
}
\description{
Generate a list of KM and Cox PH objects
}
\examples{
TBD <- c("")
}
