% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{twoGroups_DEGs}
\alias{twoGroups_DEGs}
\title{Pipeline DE Analysis Function}
\usage{
twoGroups_DEGs(
  expnData,
  clinData,
  col,
  ref,
  percent.cutoff = 0.05,
  logCPM = FALSE,
  BM = FALSE,
  GOI = NULL,
  anno = TRUE,
  ids2symbols = NULL,
  gene.name.col = "gene",
  method = "ward.D2",
  Add.Anno.Col = NULL,
  Custom.Cols = NULL,
  SkipPlots = FALSE
)
}
\arguments{
\item{expnData}{a matrix or data frame with the raw counts. Patient IDs as colnames, genes as rownames}

\item{clinData}{data.frame of clindata has patient IDs as rownames.}

\item{col}{a character string of the factor column of interest}

\item{ref}{the character strign of the reference group level (eg NBM, Neg, or control)}

\item{percent.cutoff}{the fraction (0-1 numberic) of AML samples to include when setting an expression threshold. eg 5\% of AMLs, percent=0.05.}

\item{logCPM}{boolean whether use logCPM values}

\item{BM}{boolean whether to compare group to normal bone marrows}

\item{GOI}{a character vector of genes (or numeric vector of row indices) of interest to subset at the end. keeps BH adjuted p-values more accurate.}

\item{anno}{boolean whether to add the annotation of the DEGs with TM and cell localization patterns.}

\item{ids2symbols}{ids2symbols can be NULL, NA, or a file path. NULL is for BCCA id mapping + backwards compatibility.}

\item{gene.name.col}{a character vector for column name containing gene_names if a dataframe has an alternative column name for gene symbols (eg BCL2,TP53). This CANNOT be gene_id.}

\item{method}{hierarchical clustering method}

\item{Add.Anno.Col}{character vector of additional column names from clinData to add one or two more columns for heatmap annotation color bars}

\item{Custom.Cols}{use entirely custom annotation cols}

\item{SkipPlots}{boolean whether to generate PCA, MDS, and Heatmaps}
}
\value{
list
}
\description{
Pipeline DE Analysis Function
}
\examples{

ex <- c('TBD')

}
