% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_MA_Plot_Function.r
\name{volcano_plot}
\alias{volcano_plot}
\title{function for volcano plot}
\usage{
volcano_plot(limma_eBayesFit, contrast = 1, cut.off = 2.5, label.offset = 0.5)
}
\arguments{
\item{limma_eBayesFit}{a limma eBayesFit object}

\item{contrast}{a numeric value for the contrast (column) from the coefficients to plot}

\item{cut.off}{a numeric value of log2 fold-change for gene labels on points}

\item{label.offset}{a numeric value for nudge_y in ggrepel::geom_text_repel}
}
\value{
a ggplot object
}
\description{
function for volcano plot
}
\examples{
output <- c('TBD')
}
