% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpressionPipeline_01.07.19.r
\name{voom_DE_BE}
\alias{voom_DE_BE}
\title{Voom DE with Batch effect in the Model}
\usage{
voom_DE_BE(
  expnData,
  clinData,
  col,
  percent = 0.05,
  trend = FALSE,
  logCPM = FALSE,
  normalization = FALSE,
  GOI = NULL
)
}
\arguments{
\item{expnData}{gene counts matrix}

\item{clinData}{dataframe - has patients as rownames}

\item{col}{a vector of 2 column names for co-variate/batch effect variable}

\item{percent}{percent of samples for expression threshold}

\item{trend}{boolean limma trend}

\item{logCPM}{boolean - true if input counts are already log2 CPM}

\item{normalization}{boolean to run upper quantile norm}

\item{GOI}{vector of gene names - genes of interest (GOI)}
}
\value{
list
}
\description{
Voom DE with Batch effect in the Model
}
\examples{
mat <- matrix(rnorm(10), nrow = 10, ncol = 10)
metadata <- data.frame(mutation = rep("yes","no", length.out = 10), group=rep("A","B", each = 2))
voom_DE_BE(mat, metadata, col = c("mutation","group"))
}
