% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGE_GSEA_GSVA_Function.r
\name{write.gmt}
\alias{write.gmt}
\title{Simple function to write custom list of gene-sets to gmt file format.}
\usage{
write.gmt(list_of_genesets, Description_list = NA, filename = "gene_sets.gmt")
}
\arguments{
\item{list_of_genesets}{a list of gene-sets}

\item{Description_list}{a paired list of gene-set descriptions}

\item{filename}{output filenames}
}
\value{
nothing
}
\description{
Simple function to write custom list of gene-sets to gmt file format.
}
\examples{
ex <- c("TBD")
}
