% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{colorCodes_aheatmap}
\alias{colorCodes_aheatmap}
\title{Create a color codes}
\usage{
colorCodes_aheatmap(df, random = FALSE, AML_groups = FALSE)
}
\arguments{
\item{df}{cleaned clinical characteristics, with patient IDs as rownames}

\item{random}{boolean. should colors be selected by random sampling.}
}
\value{
list with names as the input columns of the data.frame. Each list has 1 color-value mapped to 1 category (class)
}
\description{
Create a color codes
}
\examples{
ex <- c('TBD')
}
