% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmaps_Function.r
\name{colorVectors_asList}
\alias{colorVectors_asList}
\title{Create a list of color codes for all columns in a datafram}
\usage{
colorVectors_asList(df)
}
\arguments{
\item{df}{df of cleaned clinical characteristics. patient IDs as rownames}
}
\value{
list
}
\description{
Create a list of color codes for all columns in a datafram
}
\examples{
ex <- c("TBD")
}
